refactor: isolate Codex restore validation result
